Refactor urllib3/util.py into submodules
